refactor(notificacoes): split AdminNotificacoes (510→101) em hook + componentes
Extrai verbatim de src/pages/AdminNotificacoes.tsx:
- useNotificacoes: 3 queries (pendentes/histórico/stats30 com refetchInterval)
  + mutation de disparo + memos (filtros/estatísticas/chartData)
- DispatchButton, PendentesTab, HistoricoTab, StatsTab, AlertaDrawer,
  badges (SeveridadeBadge/StatusBadge), types.ts, format.ts
  (SELECT_COLUMNS/relTime/fmtDate)
- +17 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
